+++ packages/bundle/src/adaptiveCards/Attachment/CommonCard.js
@@ -7,21 +7,22 @@ import AdaptiveCardBuilder from './AdaptiveCardBuilder';
 import AdaptiveCardRenderer from './AdaptiveCardRenderer';
 import useAdaptiveCardsPackage from '../hooks/useAdaptiveCardsPackage';
 
-const { useStyleOptions } = hooks;
+const { useDirection, useStyleOptions } = hooks;
 
 const CommonCard = ({ attachment: { content } }) => {
   const [adaptiveCardsPackage] = useAdaptiveCardsPackage();
+  const [direction] = useDirection();
   const [styleOptions] = useStyleOptions();
 
   const builtCard = useMemo(() => {
     if (content) {
-      const builder = new AdaptiveCardBuilder(adaptiveCardsPackage, styleOptions);
+      const builder = new AdaptiveCardBuilder(adaptiveCardsPackage, styleOptions, direction);
 
       builder.addCommon(content);
 
       return builder.card;
     }
-  }, [adaptiveCardsPackage, content, styleOptions]);
+  }, [adaptiveCardsPackage, content, direction, styleOptions]);
 
   return <AdaptiveCardRenderer adaptiveCard={builtCard} tapAction={content && content.tap} />;
 };