+++ packages/component/src/Attachment/FileContent.js
@@ -6,6 +6,7 @@ import React from 'react';
 
 import DownloadIcon from './Assets/DownloadIcon';
 import ScreenReaderText from '../ScreenReaderText';
+import useDirection from '../hooks/useDirection';
 import useLocalize from '../hooks/useLocalize';
 import useStyleSet from '../hooks/useStyleSet';
 
@@ -25,6 +26,7 @@ const ROOT_CSS = css({
 });
 
 const FileContentBadge = ({ downloadIcon, fileName, size }) => {
+  const [direction] = useDirection();
   const formattedSize = typeof size === 'number' && format(size);
 
   return (
@@ -33,7 +35,15 @@ const FileContentBadge = ({ downloadIcon, fileName, size }) => {
         <div className="webchat__fileContent__fileName">{fileName}</div>
         {!!formattedSize && <div className="webchat__fileContent__size">{formattedSize}</div>}
       </div>
-      {downloadIcon && <DownloadIcon className="webchat__fileContent__downloadIcon" size={1.5} />}
+      {downloadIcon && (
+        <DownloadIcon
+          className={classNames(
+            'webchat__fileContent__downloadIcon',
+            direction === 'rtl' && 'webchat__fileContent__downloadIcon--rtl'
+          )}
+          size={1.5}
+        />
+      )}
     </React.Fragment>
   );
 };