+++ packages/component/src/BasicSendBox.js
@@ -13,6 +13,7 @@ import TextBox from './SendBox/TextBox';
 import TypingIndicator from './SendBox/TypingIndicator';
 import UploadButton from './SendBox/UploadButton';
 import useActivities from './hooks/useActivities';
+import useDirection from './hooks/useDirection';
 import useDictateState from './hooks/useDictateState';
 import useStyleOptions from './hooks/useStyleOptions';
 import useStyleSet from './hooks/useStyleSet';
@@ -51,12 +52,13 @@ const BasicSendBox = ({ className }) => {
   const [{ hideUploadButton }] = useStyleOptions();
   const [{ sendBox: sendBoxStyleSet }] = useStyleSet();
   const [{ SpeechRecognition } = {}] = useWebSpeechPonyfill();
+  const [direction] = useDirection();
   const [speechInterimsVisible] = useSendBoxSpeechInterimsVisible();
 
   const supportSpeechRecognition = !!SpeechRecognition;
 
   return (
-    <div className={classNames(sendBoxStyleSet + '', ROOT_CSS + '', className + '')} role="form">
+    <div className={classNames(sendBoxStyleSet + '', ROOT_CSS + '', className + '')} dir={direction} role="form">
       <TypingIndicator />
       <ConnectivityStatus />
       <SuggestedActions />