+++ packages/component/src/Styles/StyleSet/CarouselFilmStrip.js
@@ -8,15 +8,55 @@ export default function CarouselFilmStrip({ bubbleMaxWidth, bubbleMinWidth, padd
       marginBottom: -17
     },
 
-    paddingLeft: paddingRegular,
+    '&:not(.webchat__carousel--rtl)': {
+      paddingLeft: paddingRegular,
 
-    '&.webchat__carousel_indented_content > .content': {
-      marginLeft: paddingRegular
+      '&.webchat__carousel_indented_content > .content': {
+        marginLeft: paddingRegular
+      },
+
+      '& > .content': {
+        paddingRight: paddingRegular,
+
+        '& > .webchat__carousel__item_indented': {
+          marginLeft: paddingRegular
+        },
+
+        '& > ul > li:not(:last-child)': {
+          marginRight: paddingRegular
+        }
+      },
+
+      '&.webchat__carousel_extra_right_indent > .content': {
+        paddingRight: paddingRegular * 2
+      }
+    },
+
+    '&.webchat__carousel--rtl': {
+      paddingRight: paddingRegular,
+
+      '&.webchat__carousel_indented_content > .content': {
+        marginRight: paddingRegular
+      },
+
+      '& > .content': {
+        paddingLeft: paddingRegular,
+
+        '& > .webchat__carousel__item_indented': {
+          marginRight: paddingRegular
+        },
+
+        '& > ul > li:not(:last-child)': {
+          marginLeft: paddingRegular
+        }
+      },
+
+      '&.webchat__carousel_extra_right_indent > .content': {
+        paddingLeft: paddingRegular * 2
+      }
     },
 
     '& > .content': {
-      paddingRight: paddingRegular,
-
       '& > ul': {
         '&:not(:first-child)': {
           marginTop: paddingRegular
@@ -24,16 +64,8 @@ export default function CarouselFilmStrip({ bubbleMaxWidth, bubbleMinWidth, padd
 
         '& > li': {
           maxWidth: bubbleMaxWidth,
-          minWidth: bubbleMinWidth,
-
-          '&:not(:last-child)': {
-            marginRight: paddingRegular
-          }
+          minWidth: bubbleMinWidth
         }
-      },
-
-      '& > .webchat__carousel__item_indented': {
-        marginLeft: paddingRegular
       }
     }
   };