+++ packages/component/src/hooks/index.js
@@ -4,6 +4,7 @@ import useAvatarForUser from './useAvatarForUser';
 import useConnectivityStatus from './useConnectivityStatus';
 import useDictateInterims from './useDictateInterims';
 import useDictateState from './useDictateState';
+import useDirection from './useDirection';
 import useDisabled from './useDisabled';
 import useEmitTypingIndicator from './useEmitTypingIndicator';
 import useFocusSendBox from './useFocusSendBox';
@@ -55,6 +56,7 @@ export {
   useConnectivityStatus,
   useDictateInterims,
   useDictateState,
+  useDirection,
   useDisabled,
   useEmitTypingIndicator,
   useFocusSendBox,